[autoscaler v2] add test for node provider (#35593)
Why are these changes needed?
add tests for node_provider v2 and refactor the mock code
