+++ python/ray/autoscaler/v2/instance_manager/node_provider.py
@@ -1,12 +1,12 @@
 import logging
 from abc import ABCMeta, abstractmethod
-from typing import Dict, List, Set, override
+from typing import Dict, List, Set
 
 from ray.autoscaler._private.node_launcher import BaseNodeLauncher
 from ray.autoscaler.node_provider import NodeProvider as NodeProviderV1
 from ray.autoscaler.tags import TAG_RAY_USER_NODE_TYPE
 from ray.autoscaler.v2.instance_manager.config import NodeProviderConfig
-from ray.core.generated.instance_manager_pb2 import Instance, InstanceType
+from ray.core.generated.instance_manager_pb2 import Instance
 
 logger = logging.getLogger(__name__)
 
@@ -17,14 +17,14 @@ class NodeProvider(metaclass=ABCMeta):
     """
 
     @abstractmethod
-    def create_nodes(self, instance_type: InstanceType, count: int) -> List[str]:
+    def create_nodes(self, instance_type_name: str, count: int) -> List[str]:
         """Create new nodes synchronously, returns all non-terminated nodes in the cluster.
         Note that create_nodes could fail partially.
         """
         pass
 
     @abstractmethod
-    def async_terminate_nodes(self, cloud_instance_ids: List[str]) -> None:
+    def terminate_node(self, cloud_instance_id: str) -> None:
         """
         Terminate nodes asynchronously, returns immediately."""
         pass
@@ -37,7 +37,7 @@ class NodeProvider(metaclass=ABCMeta):
         pass
 
     @abstractmethod
-    def get_nodes_by_cloud_id(
+    def get_nodes_by_cloud_instance_id(
         self,
         cloud_instance_ids: List[str],
     ) -> Dict[str, Instance]:
@@ -80,37 +80,32 @@ class NodeProviderAdapter(NodeProvider):
             filtered[instance_id] = instance
         return filtered
 
-    @override
-    def create_nodes(self, instance_type: InstanceType, count: int) -> List[Instance]:
-        result = self._node_launcher.launch_node(
-            self._config.get_node_config(instance_type.name),
+    def create_nodes(self, instance_type_name: str, count: int) -> List[Instance]:
+        created_nodes = self._node_launcher.launch_node(
+            self._config.get_raw_config_mutable(),
             count,
-            instance_type.name,
+            instance_type_name,
         )
         # TODO: we should handle failures where the instance type is
         # not available
-        if result:
+        if created_nodes:
             return [
                 self._get_instance(cloud_instance_id)
-                for cloud_instance_id in result.keys()
+                for cloud_instance_id in created_nodes.keys()
             ]
         return []
 
-    @override
-    def async_terminate_nodes(self, clould_instance_ids: List[str]) -> None:
-        self._provider.terminate_node(clould_instance_ids)
+    def terminate_node(self, clould_instance_id: str) -> None:
+        self._provider.terminate_node(clould_instance_id)
 
-    @override
     def is_readonly(self) -> bool:
         return self._provider.is_readonly()
 
-    @override
     def get_non_terminated_nodes(self):
-        clould_instance_ids = self._provider.non_terminated_nodes()
-        return self.get_nodes_by_id(clould_instance_ids)
+        clould_instance_ids = self._provider.non_terminated_nodes({})
+        return self.get_nodes_by_cloud_instance_id(clould_instance_ids)
 
-    @override
-    def get_nodes_by_cloud_id(
+    def get_nodes_by_cloud_instance_id(
         self,
         cloud_instance_ids: List[str],
     ) -> Dict[str, Instance]:
@@ -123,12 +118,12 @@ class NodeProviderAdapter(NodeProvider):
         instance = Instance()
         instance.cloud_instance_id = cloud_instance_id
         if self._provider.is_running(cloud_instance_id):
-            instance.state = Instance.STARTING
+            instance.status = Instance.STARTING
         elif self._provider.is_terminated(cloud_instance_id):
-            instance.state = Instance.STOPPED
+            instance.status = Instance.STOPPED
         else:
-            instance.state = Instance.INSTANCE_STATUS_UNSPECIFIED
-        instance.interal_ip = self._provider.internal_ip(cloud_instance_id)
+            instance.status = Instance.INSTANCE_STATUS_UNSPECIFIED
+        instance.internal_ip = self._provider.internal_ip(cloud_instance_id)
         instance.external_ip = self._provider.external_ip(cloud_instance_id)
         instance.instance_type = self._provider.node_tags(cloud_instance_id)[
             TAG_RAY_USER_NODE_TYPE
